+++ lib/src/analytics/analytics_service.dart
@@ -1046,6 +1046,19 @@ class AnalyticsService {
         ),
       );
 
+  /// Sends an [AnalyticsTypes.taken] based on given [subject] and possible [parameters].
+  void taken({
+    required String subject,
+    Map<String, Object?>? parameters,
+  }) =>
+      _logAnalytic(
+        Analytic(
+          subject: subject,
+          parameters: parameters,
+          type: AnalyticsTypes.taken,
+        ),
+      );
+
   /// Sends the current screen based on given [subject] and possible [parameters].
   void screen({
     required String subject,